backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6
python-jose[cryptography]==3.3.0
httpx==0.28.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
requests==2.32.3
python-multipart==0.0.20
aiofiles==24.1.0
websockets==14.1
google-generativeai==0.8.3
